Remap keys for UltiSnip plugins
The c-tab does not work in Terminal.app for some reason. So I had to use
something else.

I also removed the <tab> to <c-p> remap because that was interfering
with using normal tabs for indention. I need to find something there
that works around that conflict. I think I saw a function in Gary
Bernhardt's vim conf that only uses the tab autocompletion when there is
no whitespace before the cursor.